work on message update
