+++ crates/api/src/systems/event_state_system.rs
@@ -515,6 +515,36 @@ impl System<Tree, StringComponentStore, RenderContext2D> for EventStateSystem {
                 .unwrap()
                 .clone();
 
+            let message_adapter = self.context_provider.message_adapter.clone();
+
+            loop {
+                if self.context_provider.message_adapter.entities().is_empty() {
+                    break;
+                }
+
+                for entity in self.context_provider.message_adapter.entities() {
+                    let mut ctx = Context::new(
+                        (entity, ecm),
+                        &theme,
+                        &self.context_provider,
+                        render_context,
+                    );
+
+                    if let Some(state) = self.context_provider.states.borrow_mut().get_mut(&entity)
+                    {
+                        state.messages(
+                            message_adapter.message_reader(entity),
+                            &mut self.registry.borrow_mut(),
+                            &mut ctx,
+                        );
+                    } else {
+                        message_adapter.remove_message_for_entity(entity);
+                    }
+
+                    drop(ctx);
+                }
+            }
+
             let mut remove_widget_list: Vec<Entity> = vec![];
 
             let mut dirty_index = 0;
